light: Refactor to add Light device
Forwarded: https://github.com/rzr/sense-hat-webthing/pull/9
